Moves update code into a separate method.
