+++ lib/helper.js
@@ -44,6 +44,13 @@ that = {
     }
     return dest;
   },
+
+  // Determines whether a value is an operator (i.e., starts with a $ sign).
+  // Assumes the value is a string.
+  isOperator: function(value) {
+    return value.length > 0 && value[0] === '$';
+  },
+
   isEmpty: function (obj) {
     for (var prop in obj) {
       if (obj.hasOwnProperty(prop)) {
